+++ src/main/java/dev/gonzalotorrestz/onlinevotingsystem/constraints/PasswordConstraintValidator.java
@@ -1,0 +1,53 @@
+package dev.gonzalotorrestz.onlinevotingsystem.constraints;
+
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+import org.passay.*;
+import org.springframework.stereotype.Component;
+
+import java.util.Arrays;
+import java.util.List;
+
+@Component
+public class PasswordConstraintValidator implements ConstraintValidator<ValidPassword, String> {
+
+    private final PasswordValidator passwordValidator;
+
+    public PasswordConstraintValidator() {
+        passwordValidator = new PasswordValidator(getRules());
+    }
+
+    @Override
+    public void initialize(ValidPassword arg0) {
+    }
+
+    @Override
+    public boolean isValid(String password, ConstraintValidatorContext context) {
+        PasswordValidator validator = new PasswordValidator(getRules());
+
+        RuleResult result = validator.validate(new PasswordData(password));
+        if (result.isValid()) {
+            return true;
+        }
+        List<String> messages = validator.getMessages(result);
+
+        String messageTemplate = String.join(" ", messages);
+        context.buildConstraintViolationWithTemplate(messageTemplate)
+                .addConstraintViolation()
+                .disableDefaultConstraintViolation();
+        return false;
+    }
+
+    public List<Rule> getRules() {
+        return Arrays.asList(
+                new LengthRule(8, 30),
+                new CharacterRule(EnglishCharacterData.UpperCase, 1),
+                new CharacterRule(EnglishCharacterData.LowerCase, 1),
+                new CharacterRule(EnglishCharacterData.Digit, 1),
+                new CharacterRule(EnglishCharacterData.Special, 1),
+                new IllegalSequenceRule(EnglishSequenceData.Alphabetical, 3, false),
+                new IllegalSequenceRule(EnglishSequenceData.Numerical, 3, false),
+                new IllegalSequenceRule(EnglishSequenceData.USQwerty, 5, false),
+                new WhitespaceRule());
+    }
+}