+++ src/main/java/dev/gonzalotorrestz/onlinevotingsystem/service/UserService.java
@@ -1,13 +1,16 @@
 package dev.gonzalotorrestz.onlinevotingsystem.service;
 
+import dev.gonzalotorrestz.onlinevotingsystem.dto.RegisteredUserDTO;
 import dev.gonzalotorrestz.onlinevotingsystem.dto.UserDTO;
 import dev.gonzalotorrestz.onlinevotingsystem.model.User;
 import dev.gonzalotorrestz.onlinevotingsystem.repository.UserRepository;
 import dev.gonzalotorrestz.onlinevotingsystem.util.UserEntityDTOConverter;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -15,25 +18,22 @@ import java.util.List;
 public class UserService {
     private final UserRepository userRepository;
     private final UserEntityDTOConverter userConverter;
+    private final BCryptPasswordEncoder passwordEncoder;
 
-    public UserService(UserRepository userRepository, UserEntityDTOConverter userConverter)
-    {
-            this.userRepository = userRepository;
-            this.userConverter = userConverter;
+    public UserService(UserRepository userRepository, UserEntityDTOConverter userConverter, BCryptPasswordEncoder passwordEncoder) {
+        this.userRepository = userRepository;
+        this.userConverter = userConverter;
+        this.passwordEncoder = passwordEncoder;
     }
 
-    public UserDTO createUser(UserDTO userDTO) {
-        User user = userConverter.convertToEntity(userDTO);
-        User savedUser = userRepository.save(user);
-        return userConverter.convertToDTO(savedUser);
+    public RegisteredUserDTO registerUser(@NotNull UserDTO userDTO) {
+        userDTO.setRegistrationDate(LocalDateTime.now());
+        userDTO.setPassword(passwordEncoder.encode(userDTO.getPassword()));
+        User savedUser = userRepository.save(userConverter.convertToEntity(userDTO));
+        return userConverter.convertToRegisteredUserDTO(savedUser);
     }
 
-    public UserDTO getUserByUsername(String username) {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new EntityNotFoundException("User not found with username: " + username));
-        return userConverter.convertToDTO(user);
-    }
-    public List<UserDTO> getAllUsers(){
+    public List<UserDTO> getAllUsers() {
         var userList = userRepository.findAll();
         return userConverter.convertToDTOList(userList);
     }