+++ src/main/java/dev/gonzalotorrestz/onlinevotingsystem/util/EntityDTOConverter.java
@@ -2,5 +2,6 @@ package dev.gonzalotorrestz.onlinevotingsystem.util;
 
 public interface EntityDTOConverter<Entity, DTO> {
     DTO convertToDTO(Entity entity);
+
     Entity convertToEntity(DTO dto);
 }